collect/requirements.txt
python-magic==0.4.27
PyYAML==6.0.3
scapy==2.6.1
